修改了添加地址的显示bug和删除地址的后台逻辑

+++ shop/src/com/shopping/web/action/DeleteAddressAction.java
@@ -8,7 +8,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.shopping.dao.AddressDao;
+import com.shopping.dao.OrderDao;
 import com.shopping.dao.daoImpl.AddressDaoImpl;
+import com.shopping.dao.daoImpl.OrderDaoImpl;
 import com.shopping.db.DBHelper;
 import com.shopping.pojo.Address;
 
@@ -24,19 +26,27 @@ public class DeleteAddressAction extends ActionFather{
   		
   		//获得要删除的地址的id
   		String id = request.getParameter("deleteId");
-  		int returnInfo = 0;
+  		int returnInfo = -1;
   		AddressDao ad = new AddressDaoImpl();
+  		OrderDao od = new OrderDaoImpl();
   		Connection conn = DBHelper.getConnection();
   		if(id!=null && !id.equals("") && !id.equals("null")){
   			int deleteId = Integer.parseInt(id);
+			//查询地址是否被用于订单表
+			boolean used = od.selectByAddressId(deleteId, conn);
   			Address add = new Address();
   			add.setId(deleteId);
-  			boolean b = ad.deleteById(add, conn);
-  			if(b==true){
-  	  			returnInfo = 1;
-  			}
+			if(used==true){
+				add.setMemberId("");
+				//将对应id的memberId设置为空
+				ad.update("updateMemberIdNull", add, conn);
+			}else{
+	  			boolean b = ad.deleteById(add, conn);
+	  			if(b==true){
+	  	  			returnInfo = 1;
+	  			}
+			}
   		}
-
   		PrintWriter out = null;
 		try {
 			out = response.getWriter();
